+++ XFMG/Controller/Media.php
@@ -3,8 +3,8 @@
 namespace Xfrocks\Api\XFMG\Controller;
 
 
+use AbstractController;
 use XF\Mvc\ParameterBag;
-use Xfrocks\Api\Controller\AbstractController;
 use Xfrocks\Api\Util\PageNav;
 
 class Media extends AbstractController
@@ -46,7 +46,7 @@ class Media extends AbstractController
 
     protected function actionSingle($itemId)
     {
-        $item = $this->assertViewableItem($itemId);
+        $item = $this->assertViewableMediaItem($itemId);
 
         $data = [
             'item' => $this->transformEntityLazily($item)
@@ -111,7 +111,7 @@ class Media extends AbstractController
 
     public function actionPutIndex(ParameterBag $params)
     {
-        $item = $this->assertViewableItem($params->media_id);
+        $item = $this->assertViewableMediaItem($params->media_id);
         if (!$item->canEdit($error)) {
             return $this->noPermission($error);
         }
@@ -137,7 +137,7 @@ class Media extends AbstractController
 
     public function actionDeleteIndex(ParameterBag $params)
     {
-        $item = $this->assertViewableItem($params->media_id);
+        $item = $this->assertViewableMediaItem($params->media_id);
         if (!$item->canDelete('soft', $error))
         {
             return $this->noPermission($error);
@@ -148,74 +148,5 @@ class Media extends AbstractController
         $deleter->delete('soft');
 
         return $this->actionSingle($item->media_id);
-    }
-
-    /**
-     * @param int $albumId
-     * @param array $extraWith
-     * @return \XFMG\Entity\Album
-     * @throws \XF\Mvc\Reply\Exception
-     */
-    protected function assertViewableAlbum($albumId, array $extraWith = [])
-    {
-        /** @var \XFMG\Entity\Album $album */
-        $album = $this->assertRecordExists(
-            'XFMG:Album',
-            $albumId,
-            $extraWith,
-            'xfmg_requested_album_not_found'
-        );
-
-        if (!$album->canView($error)) {
-            throw $this->exception($this->noPermission($error));
-        }
-
-        return $album;
-    }
-
-    /**
-     * @param int $categoryId
-     * @param array $extraWith
-     * @return \XFMG\Entity\Category
-     * @throws \XF\Mvc\Reply\Exception
-     */
-    protected function assertViewableCategory($categoryId, array $extraWith = [])
-    {
-        /** @var \XFMG\Entity\Category $category */
-        $category = $this->assertRecordExists(
-            'XFMG:Category',
-            $categoryId,
-            $extraWith,
-            'xfmg_requested_category_not_found'
-        );
-
-        if (!$category->canView($error)) {
-            throw $this->exception($this->noPermission($error));
-        }
-
-        return $category;
-    }
-
-    /**
-     * @param int $mediaId
-     * @param array $extraWith
-     * @return \XFMG\Entity\MediaItem
-     * @throws \XF\Mvc\Reply\Exception
-     */
-    protected function assertViewableItem($mediaId, array $extraWith = [])
-    {
-        /** @var \XFMG\Entity\MediaItem $item */
-        $item = $this->assertRecordExists(
-            'XFMG:MediaItem',
-            $mediaId,
-            $extraWith,
-            'xfmg_requested_media_item_not_found'
-        );
-
-        if (!$item->canView($error)) {
-            throw $this->exception($this->noPermission($error));
-        }
-
-        return $item;
     }
 }
